yl - finished Forums

+++ src/pages/Forum.js
@@ -3,29 +3,73 @@ import React, {Component} from 'react';
 import axios from 'axios';
 import Container from "react-bootstrap/Container";
 import getUser from "../utils/get-user";
-import { threads } from "./data";
-import { ThreeSixtySharp } from "@material-ui/icons";
 
 //const textStyle = {maxWidth: "100%", width: "700px"}
 
+
+//style for posts
+const styles = {
+  postStyle: {
+    borderStyle: 'outset',
+    marginTop: "5px",
+    marginBottom: "5px",
+    textAlign: 'left',
+    padding: '5px',
+    wordBreak: 'break-all',
+    overflowWrap: 'break-word',
+  },
+  //style for post list
+  postListStyle: {
+    padding: '0',
+  },
+  postLinkStyle: {
+    color: 'inherit',    
+  },
+  categoryStyle: {
+    borderRadius: '20px',
+    padding: '5px',
+    paddingLeft: '10px',
+    paddingRight: '10px',
+    fontSize: '12px',
+    textDecoration: 'none',
+    background: 'blue',
+    color: "white",
+  },
+  statusStyle: {
+    borderRadius: '20px',
+    padding: '5px',
+    paddingLeft: '10px',
+    paddingRight: '10px',
+    fontSize: '12px',
+    textDecoration: 'none',
+    background: 'red',
+    color: "white",
+  },
+  delEditButtons:{
+    fontSize: '13px',
+    float: 'left',
+  }
+}
 // this is the main page of forum, see list of posts
 
 // Post Component
 const Post = props => (
-  <li className="row">
-    <a href={`/forum/${props.post._id}`}>
-      <h4 className="title">
-        {props.post.title} <small>{props.post.category} { props.post.status == 'CLOSED' && ' - ' + props.post.status}</small> 
-      </h4>
-      <div className="bottom">
-        <p className="info-line">
-            <span className="author">{props.post.username}</span> - <span className="date">{props.post.date.substring(0,10)}</span> - <span className="comment-count">{props.post.numComments} comments</span>
-        </p>
-      </div>
-    </a>
-    <p className="buttons" style={{"fontSize":"10px"}}>
-        <input type="button" value="Delete Post" style={{float: "center"}} onClick={() => {props.deletePost()}}/> <input type="button" value="Edit Post" style={{float: "center"}} onClick={() => {props.editPost()}}/> <input type="button" value="Change Status" style={{float: "center"}} onClick={() => {props.changeStatus()}}/>
-    </p>
+  <li className="row" style={styles.postStyle}>
+    <body>
+      <a href={`/forum/${props.post._id}`} style={styles.postLinkStyle}>
+        <h4 className="title">
+          {props.post.title}  <small style={styles.categoryStyle}>{props.post.category}</small>  <small >{ props.post.status === 'CLOSED' &&  <span style={styles.statusStyle}>{props.post.status}</span>}</small> 
+        </h4>
+        <div className="bottom">
+          <p className="info-line">
+              <span className="author">{props.post.username}</span> - <span className="date">{props.post.date.substring(0,10)}</span> - <span className="comment-count">{props.post.numComments} comments</span>
+          </p>
+        </div>
+      </a>
+      <p className="buttons" style={styles.delEditButtons}>
+        <input type="button" value="Change Status" onClick={() => {props.changeStatus()}}/> <input type="button" value="Edit" onClick={() => {props.editPost()}}/> <input type="button" value="Delete" onClick={() => {props.deletePost()}}/>
+      </p>
+    </body>
   </li>
 )
 
@@ -46,6 +90,7 @@ export default class ForumPost extends Component{
     console.log(this.state.user);
     axios.get('http://localhost:3001/posts') //get request
          .then(res=>{
+           console.log(res.data)
            this.setState({posts: res.data}) //sets posts array to db array
          }) 
          .catch(err => {
@@ -58,24 +103,27 @@ export default class ForumPost extends Component{
     says 'Post deleted!'
     */
   deletePost(id){
+    if (window.confirm("Are you sure you want to delete this post?")) {
       axios.delete(`http://localhost:3001/posts/${id}`)
-           .then(res => {
-               console.log(res.data)
-               //update post
-               axios.get('http://localhost:3001/posts') //get request
+          .then(res => {
+              console.log(res.data)
+              //update post
+              axios.get('http://localhost:3001/posts') //get request
                     .then(res=>{
                       this.setState({posts: res.data}) //sets posts array to db array
                     }) 
                     .catch(err => {
                       console.log(err);
                     })
-           })
-           .catch(err =>{
-               console.log('Error: ' + err);
-           })
+          })
+          .catch(err =>{
+              console.log('Error: ' + err);
+          })
       
       console.log("delPost");
+      window.alert("Post Deleted!");
       window.location = '/forum';
+    }
   }
 
   /* edit the post: redirect to createPost page with every info in to resumbit,, after
@@ -213,6 +261,24 @@ export default class ForumPost extends Component{
                 console.log(err);
               })
       }
+      else if(this.state.filter === "Open Posts"){
+        axios.get('http://localhost:3001/posts/Open-Posts') //get request
+              .then(res=>{
+                this.setState({posts: res.data}) //sets posts array to db array
+              }) 
+              .catch(err => {
+                console.log(err);
+              })
+      }
+      else if(this.state.filter === "Closed Posts"){
+        axios.get('http://localhost:3001/posts/Closed-Posts') //get request
+              .then(res=>{
+                this.setState({posts: res.data}) //sets posts array to db array
+              }) 
+              .catch(err => {
+                console.log(err);
+              })
+      }
     })
 
   }
@@ -222,6 +288,7 @@ export default class ForumPost extends Component{
       <Layout user={this.state.user}>
         <Container>
           <h1><a href="/forum" style={{"textDecoration": "none", "color":"inherit"}}>Bike Forum</a></h1>
+          <hr/>
           <br></br>
           <div className="main">
             {/* Create new post (should only be for users) */}
@@ -230,7 +297,7 @@ export default class ForumPost extends Component{
               { this.state.user && <input type="button" value="Create New Post" onClick={() => {window.location = "forum/create-post"}}/>}  
               { this.state.user && <br></br>}       
             </div> } 
-
+            <br/>
             <select ref="categoryInput" 
                         required className="form-control" 
                         value={this.state.filter} 
@@ -241,12 +308,14 @@ export default class ForumPost extends Component{
               <option>Lost and Found</option>
               <option>Crash Reports</option>
               <option>Others</option>
+              <option>Open Posts</option>
+              <option>Closed Posts</option>
 
             </select>
             <br/>
             {/* <pre>{JSON.stringify(this.state.user, null, 2)}</pre> */}
           
-            <ol>
+            <ol style={styles.postListStyle}>
               {this.postList()} {/*rendering all of container list elements*/}
             </ol>
           </div>
